requirements.txt
flask==3.1.0
flask_sqlalchemy==3.1.1
PyJWT==2.10.1
mysqlclient==2.2.6
uwsgi==2.0.28
cryptography==44.0.0
#waitress==3.0.2